+++ circuit1.2.py
@@ -1,19 +1,6 @@
 from bloqade import move
 from iquhack_scoring import MoveScorer
-import math
-import matplotlib.pyplot as plt
-
-pi = math.pi
-
-@move.vmove()
-def rx(state: move.core.AtomState, indices, theta) -> move.core.AtomState:
-    state = move.LocalXY(atom_state=state,x_exponent=theta,axis_phase_exponent=0,indices=indices)
-    return state
-
-@move.vmove()
-def ry(state: move.core.AtomState, indices, theta) -> move.core.AtomState:
-    state = move.LocalXY(atom_state=state,x_exponent=theta,axis_phase_exponent=-pi/2,indices=indices)
-    return state
+from utils import pi, rx, ry, make_gif
 
 @move.vmove()
 def circuit1():
@@ -59,6 +46,4 @@ ccx q[0],q[1],q[2];
 scorer = MoveScorer(circuit1, expected_qasm)
 print(scorer.score())
 
-ani = scorer.animate()
-ani.save("circuit1.2.gif")
-plt.show()
+make_gif(scorer, "circuit1.2.gif")
